Extract the customer-portal contracts into a publishable leaf
Platform's portal needs to validate /v1/public/customer-portal payloads in
a browser. It could only get those schemas from @voyant-travel/public-api,
the server composition layer that depends on eleven domain modules — so
consuming them meant a 71-package closure, 59 of it private, i.e. undoing
voyant#4098.

The schemas never needed any of that. All 664 lines of them import zod and
one 49-line helper from @voyant-travel/schema-kit, which is already on the
public surface. They are now @voyant-travel/public-api-contracts, closure
of 2, nothing private.

public-api keeps its ./customer-portal/public-validation subpath as a
re-export so every in-repo consumer still imports them from the module that
owns the surface they describe. public-api-react drops its dependency on
the server package entirely.

Not inlined to reach literally zero dependencies: schema-kit/email exists
because zod's default email regex uses lookaround, which RE2-style LLM
schema validators reject outright (voyant#4598), and
verify:tool-schema-portability fails the build if such a pattern reaches a
Tool schema. Copying that pattern would duplicate a load-bearing invariant
for no gain, since schema-kit is public already.

verify:public-surface: 15 packages, closure 15 — closed. public-api 213
passed, public-api-react 33 passed, contracts build and typecheck clean.
