+++ packages/public-api-contracts/package.json
@@ -1,0 +1,64 @@
+{
+  "name": "@voyant-travel/public-api-contracts",
+  "version": "0.0.0",
+  "description": "Customer-portal request and response contracts for the Public API — validate /v1/public/customer-portal payloads in a browser without the server.",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/voyant-travel/voyant.git",
+    "directory": "packages/public-api-contracts"
+  },
+  "type": "module",
+  "sideEffects": false,
+  "exports": {
+    ".": "./src/index.ts",
+    "./bookings": "./src/bookings.ts",
+    "./common": "./src/common.ts",
+    "./profile": "./src/profile.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc -p tsconfig.typecheck.json",
+    "lint": "biome check src/",
+    "test": "vitest run --passWithNoTests",
+    "build": "tsc -p tsconfig.build.json",
+    "clean": "rm -rf dist tsconfig.tsbuildinfo",
+    "prepack": "pnpm run build"
+  },
+  "files": [
+    "dist"
+  ],
+  "publishConfig": {
+    "access": "public",
+    "exports": {
+      ".": {
+        "types": "./dist/index.d.ts",
+        "import": "./dist/index.js",
+        "default": "./dist/index.js"
+      },
+      "./bookings": {
+        "types": "./dist/bookings.d.ts",
+        "import": "./dist/bookings.js",
+        "default": "./dist/bookings.js"
+      },
+      "./common": {
+        "types": "./dist/common.d.ts",
+        "import": "./dist/common.js",
+        "default": "./dist/common.js"
+      },
+      "./profile": {
+        "types": "./dist/profile.d.ts",
+        "import": "./dist/profile.js",
+        "default": "./dist/profile.js"
+      }
+    }
+  },
+  "dependencies": {
+    "@voyant-travel/schema-kit": "workspace:^",
+    "zod": "catalog:"
+  },
+  "devDependencies": {
+    "@voyant-travel/voyant-typescript-config": "workspace:^",
+    "typescript": "catalog:",
+    "vitest": "catalog:"
+  }
+}
