+++ packages/public-api-contracts/src/bookings.ts
