+++ packages/public-api-contracts/src/common.ts
