+++ packages/public-api-contracts/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./bookings.js"
+export * from "./profile.js"
