+++ packages/public-api-react/package.json
@@ -75,6 +75,7 @@
   },
   "dependencies": {
     "@voyant-travel/i18n": "workspace:^",
+    "@voyant-travel/public-api-contracts": "workspace:^",
     "@voyant-travel/react": "workspace:^"
   },
   "devDependencies": {
